Revert "Allow Horde_Form_Type_boolean to accept hidden true or false values"
This reverts commit 5e1ee9cda6b078839da2a24348b8198b5c8a9064.

This breaks Horde_Form being able to uncheck a checkbox that is currently
checked since $var->getValue() returns the "default" value if the value is
not sent in the form post. For boolean types, this is a obviously a boolean value
so is_boolean will ALWAYS be true in this case.

This fixes, e.g., not being able to unset boolean type permissions in the
permissios UI.