Loosen the the test for the date format (Bug #14130).
Theoretically it might have been intended to explicitly pass an empty string (as opposed to a null value) as the date format. But this doesn't make sense because that would generate a form field that never has a result value. Beside that we already *do* catch an empty value, even if it's a strict null check.