Workaround hidden read-only fields not being detected during form submission.
Setting hidden fields to read-only doesn't make any sense because they are always read-only. The problem is that read-only fields are not set as expected when submitting forms and thus their values no longer considered since the recent changes. This breaks forms e.g. in Whups that had hidden fields set to read-only and required at the same time. Workaround this to maintain BC until this is fixed properly.